rename NftPurchase to NftPurchaseBottomDrawer, use BottomDrawer
