+++ packages/atlas/src/views/global/NftPurchaseBottomDrawer/index.ts
@@ -1,0 +1,1 @@
+export * from './NftPurchaseBottomDrawer'
