--- packages/atlas/src/views/viewer/NftPurchaseView/index.ts
@@ -1,1 +1,0 @@
-export * from './NftPurchaseView'
